// store/application/types/enums/action-type.ts
/**
 * Actions supported by application store module
 */
export enum ActionType {
    GET_IMAGE_URL_TEMPLATES = "GET_IMAGE_URL_TEMPLATES",
    ADD_SELECTED_ID_PROGRAM = "ADD_SELECTED_ID_PROGRAM",
    FETCH_PRIVACY_TEMPLATE = "FETCH_PRIVACY_TEMPLATE",
    FETCH_ABOUT_TEMPLATE = "FETCH_ABOUT_TEMPLATE",
    SAVE_VISIBLE_DESCRIPTION = "SAVE_VISIBLE_DESCRIPTION",
    FETCH_FORGOT_LINK = "FETCH_FORGOT_LINK",
}